Reading JSON file and Creating 'Book' object
I put it in constructor for now. The main() method in Book class was created for testing only, we will need to implement its contents to methods later on. I also added a JSON file in "books" directory named "test.json". Right now GitHub Desktop says there are 37 changed files and I do not know the reason, we may need to change something with possibly .gitignore